+++ test/plugins.js
@@ -1,0 +1,9 @@
+
+/*
+  Placeholder test file for plugins. Ensures they show up as bleeding red
+  in the coverage report :)
+*/
+
+import '../src/plugins/toReactShape'
+import '../src/plugins/koaMiddleware'
+import '../src/plugins/connectMiddleware'